# gh-90102: Remove isatty call during regular open