--- a/Lib/_pyio.py
+++ b/Lib/_pyio.py
@@ -238,7 +238,7 @@ def open(file, mode="r", buffering=-1, encoding=None, errors=None,
     result = raw
     try:
         line_buffering = False
-        if buffering == 1 or buffering < 0 and raw._isatty_openonly():
+        if buffering == 1 or buffering < 0 and raw._isatty_open_only():
             buffering = -1
             line_buffering = True
         if buffering < 0:
@@ -1794,14 +1794,15 @@ class FileIO(RawIOBase):
         self._checkClosed()
         return os.isatty(self._fd)
 
-    def _isatty_openonly(self):
+    def _isatty_open_only(self):
         """Checks whether the file is a TTY using an open-only optimization.
 
-        Normally isatty always makes a system call. In the case of open() there
-        is a _inside the same python call_ stat result which we can use to
-        skip that system call for non-character files. Outside of that context
-        this is subject to TOCTOU issues (the FD has been returned to user code
-        and arbitrary syscalls could have happened).
+        TTYs are always character devices. If the interpreter knows a file is
+        not a character device when it would call `isatty`, can skip that call.
+        Inside `open()`  there is a fresh stat result that contains that
+        information. Use the stat result to skip a system call. Outside of that
+        context TOCTOU issues (the fd could be arbitrarily modified by
+        surrounding code).
         """
         if (self._stat_atopen is not None
             and not stat.S_ISCHR(self._stat_atopen.st_mode)):