--- a/Modules/_io/fileio.c
+++ b/Modules/_io/fileio.c
@@ -1206,6 +1206,7 @@ _io_FileIO_isatty_impl(fileio *self)
 }
 
 /* Checks whether the file is a TTY using an open-only optimization.
+
    Normally isatty always makes a system call. In the case of open() there
    is a _inside the same python call_ stat result which we can use to
    skip that system call for non-character files. Outside of that context