--- a/Modules/_io/fileio.c
+++ b/Modules/_io/fileio.c
@@ -1208,8 +1208,8 @@ _io_FileIO_isatty_impl(fileio *self)
 /* Checks whether the file is a TTY using an open-only optimization.
 
    TTYs are always character devices. If the interpreter knows a file is
-   not a character device when it would call `isatty`, can skip that call.
-   Inside `open()`  there is a fresh stat result that contains that
+   not a character device when it would call ``isatty``, can skip that
+   call. Inside ``open()``  there is a fresh stat result that contains that
    information. Use the stat result to skip a system call. Outside of that
    context TOCTOU issues (the fd could be arbitrarily modified by
    surrounding code). */