@vstinner 2024-10-08
Merged, thank you. At the first read, I didn't understand the relationship between S_ISCHR() and isatty(). But your comment makes sense and explains it correctly.